simple update implementation

+++ src/updater.ts
@@ -2,15 +2,10 @@ import CustomerModel, { ICustomer } from "./models/Customer";
 import ModeModel, { IStatus } from "./models/ModeStatus";
 import mongoose, { Schema } from "mongoose";
 import AnonymizeCustomerModel from "./models/CustomerAnonymised";
-import QueueModel from "./models/Queue";
 
 class Updater {
-  // customers ready to update
-  private customersMap: Map<Schema.Types.ObjectId, ICustomer> = new Map<
-    Schema.Types.ObjectId,
-    ICustomer
-  >();
   static upsertPackSize = 1000;
+
   async isFullReindexModeRunning(): Promise<boolean> {
     const mode: IStatus | null = await ModeModel.findOne();
     const result = mode ? Boolean(mode.fullReindex) : true;
@@ -52,65 +47,53 @@ class Updater {
     return anonymizedCustomer;
   }
 
-  //accept new customers from events
-  pushCustomers(customers: ICustomer[]) {
-    if (this.customersMap.size < Updater.upsertPackSize) {
-      //not more than 1000 in one pack to insert, rest must be persisted in toUpdateIds collection
-      customers
-        .slice(0, Updater.upsertPackSize - this.customersMap.size)
-        .forEach((customer) => {
-          if (customer._id) {
-            this.customersMap.set(
-              customer._id,
-              this.anonymizeCustomerData(customer)
-            );
-          }
-        });
-    }
-  }
-
-  private async getCustomersFromQueue(
-    limit: number,
-    omitIds: (Schema.Types.ObjectId | undefined)[]
-  ): Promise<ICustomer[]> {
-    const ids = await Queue.find({ id: { $nin: omitIds } }, { id: true })
-      .limit(limit)
-      .distinct("id");
-    return CustomerModel.find({ _id: { $in: ids } });
-  }
-  async addCustomersFromQueue(omitIds: (Schema.Types.ObjectId | undefined)[]) {
-    if (this.customersMap.size < Updater.upsertPackSize) {
-      const addFromQueueCount = Updater.upsertPackSize - this.customersMap.size;
-      const customersFromQueue = await this.getCustomersFromQueue(
-        addFromQueueCount,
-        omitIds
-      );
-      this.pushCustomers(customersFromQueue);
-    }
-  }
-
   async upsertCustomers(): Promise<void> {
     if (await this.isFullReindexModeRunning()) return;
     const session = await mongoose.startSession();
     try {
       console.log("START TRANSACTION");
       session.startTransaction();
-      const omitIds = Array.from(this.customersMap.keys());
-      await this.addCustomersFromQueue(omitIds);
-      const upsertQuery = Array.from(this.customersMap.values()).map((customer) => ({
-        updateOne: {
-          filter: { _id: customer._id },
-          update: { $set: customer },
-          upsert: true,
-        },
-      }));
+      const helpInfo: IStatus | null = await ModeModel.findOne({});
+      if (!helpInfo) {
+        console.error("mode collection is broken!");
+        return;
+      }
+      const { lastUpdateTime } = helpInfo;
+      const customersToUpdate: ICustomer[] = await CustomerModel.find({
+        updatedAt: { $gt: lastUpdateTime },
+      })
+        .sort({ updatedAt: 1 })
+        .limit(Updater.upsertPackSize);
 
-      const ids = Array.from(this.customersMap.keys());
-      await AnonymizeCustomerModel.bulkWrite(upsertQuery, { session });
-      await Queue.deleteMany({ id: { $in: ids } }).session(session);
+      if (!customersToUpdate.length) {
+        await session.commitTransaction();
+        await session.endSession();
+        return;
+      }
+      const newLastUpdateTime =
+        customersToUpdate[customersToUpdate.length - 1].updatedAt;
+      const query = customersToUpdate.map((customer) => {
+        const anonymizedCustomer = this.anonymizeCustomerData(customer);
+        return {
+          updateOne: {
+            filter: {
+              _id: customer._id,
+            },
+            update: {
+              $set: anonymizedCustomer,
+            },
+            upsert: true,
+          },
+        };
+      });
+
+      await AnonymizeCustomerModel.bulkWrite(query, { session });
+      await ModeModel.findOneAndUpdate(
+        {},
+        { lastUpdateTime: newLastUpdateTime }
+      ).session(session);
       await session.commitTransaction();
       await session.endSession();
-      this.customersMap = new Map<Schema.Types.ObjectId, ICustomer>();
       console.log("END TRANSACTION");
     } catch (e) {
       console.error("ERROR DURING TRANSACTION");
@@ -123,24 +106,34 @@ class Updater {
   async fullReindex(): Promise<void> {
     //todo add transaction
     await ModeModel.deleteMany();
-    await new ModeModel({ fullReindex: true }).save();
-    await QueueModel.deleteMany();
+    await new ModeModel({
+      fullReindex: true,
+      lastUpdateTime: new Date(),
+    }).save();
     await AnonymizeCustomerModel.deleteMany();
+    let lastUpdateTime : Date;
     const updatedIds: Schema.Types.ObjectId[] = [];
     const customersCount = await CustomerModel.count();
     while (updatedIds.length < customersCount) {
       const toUpdate: ICustomer[] = await CustomerModel.find({
         _id: { $nin: updatedIds },
-      }).limit(Updater.upsertPackSize);
+      }).limit(Updater.upsertPackSize).sort({ updatedAt: 1 });
+
+
       const anonymizedToUpdate = toUpdate.map((customer) => {
         if (customer._id) updatedIds.push(customer._id);
         return this.anonymizeCustomerData(customer);
       });
       await AnonymizeCustomerModel.insertMany(anonymizedToUpdate);
+      if(updatedIds.length === customersCount) {
+        lastUpdateTime = toUpdate[toUpdate.length - 1].updatedAt;
+        await ModeModel.deleteMany();
+        await new ModeModel({
+          fullReindex: false,
+          lastUpdateTime: lastUpdateTime,
+        }).save();
+      }
     }
-
-    await ModeModel.deleteMany();
-    await new ModeModel({ fullReindex: false }).save();
   }
 }
